move `Var`/`Lit`/`Pol` into separate `lit` crate
These types have plenty of use cases that don't depend on specific IR
data structures, so it should be possible to depend on them without
pulling in the `ir` crate as dependency.
