+++ table_seq/src/lib.rs
@@ -108,12 +108,12 @@ use hashbrown::HashTable;
 
 mod node_allocator;
 
-pub mod table_seq;
 pub mod set_seq;
+pub mod table_seq;
 
 #[doc(inline)]
-pub use table_seq::TableSeq;
-#[doc(inline)]
 pub use set_seq::SetSeq;
+#[doc(inline)]
+pub use table_seq::TableSeq;
 
 // TODO Implement MapSeq on top of TableSeq
